requirements.txt
absl-py==2.1.0
addict==2.4.0
affine==2.4.0
aiohttp==3.9.1
aiomultiprocess==0.9.0
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.2
annotated-types==0.6.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
bs4==0.0.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
comm==0.2.1
contextily==1.6.0
contourpy==1.2.0
coord-convert==0.2.1
coverage==7.5.3
crcmod==1.7
cryptography==42.0.5
cycler==0.12.1
datasets==2.18.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
dnspython==2.4.2
esdk-obs-python==3.23.12
exceptiongroup==1.2.0
executing==2.0.1
Faker==23.2.1
fastjsonschema==2.19.1
filelock==3.14.0
fiona==1.9.6
fonttools==4.50.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.5.0
ftfy==6.2.0
gast==0.5.4
generate-od==0.1
geographiclib==2.0
geojson==3.1.0
geopandas==0.14.4
geopy==2.4.1
grpcio==1.64.0
grpcio-tools==1.62.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
igraph==0.11.5
importlib_metadata==7.1.0
iniconfig==2.0.0
ipykernel==6.28.0
ipython==8.20.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.3.2
json5==0.9.14
jsonlines==4.0.0
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.4
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
Levenshtein==0.25.1
lru-dict==1.3.0
lxml==5.2.2
Markdown==3.6
MarkupSafe==2.1.3
matplotlib==3.8.3
matplotlib-inline==0.1.6
mercantile==1.2.1
mistune==3.0.2
modelscope==1.13.3
mosstool==0.2.5
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.16
nbclient==0.9.0
nbconvert==7.14.1
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.3
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
open-clip-torch==2.24.0
openai==1.25.1
oss2==2.18.4
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.5.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.0.0
pyarrow-hotfix==0.6
pycityproto==1.13.1
pycitysim==1.13.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.8.0
Pygments==2.17.2
pymongo==4.7.2
pyparsing==3.1.2
pyproj==3.6.1
PySocks==1.7.1
pytest==8.2.1
pytest-cov==5.0.0
pytest-cover==3.0.0
pytest-coverage==0.0
python-dateutil==2.8.2
python-json-logger==2.0.7
python-moss==0.2.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
rapidfuzz==3.9.3
rasterio==1.3.9
referencing==0.32.1
regex==2023.12.25
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.16.2
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.0
seaborn==0.13.2
Send2Trash==1.8.2
sentencepiece==0.2.0
setproctitle==1.3.3
shapely==2.0.4
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
stringcase==1.2.0
sympy==1.12.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
terminado==0.18.0
texttable==1.7.0
thefuzz==0.22.1
threadpoolctl==3.5.0
timm==1.0.3
tinycss2==1.2.1
tomli==2.0.1
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
triton==2.3.0
types-protobuf==4.24.0.20240106
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
uri-template==1.3.0
urllib3==2.1.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
xxhash==3.4.1
xyzservices==2024.4.0
yapf==0.40.2
yarl==1.9.4
zipp==3.18.1
mosstool==0.2.5
